backend/requirements.txt
flask
flask_cors
torch
faiss-cpu
Pillow
pandas
tqdm
git+https://github.com/openai/CLIP.git